+++ src/pages/details.tsx
@@ -3,14 +3,14 @@ import { GetServerSideProps } from "next"
 import { useRouter } from "next/router"
 import { Session, getServerSession } from "next-auth"
 
-import { getCompanyInfo } from "@/lib/qbo-api"
 import { authOptions } from "./api/auth/[...nextauth]"
-import { Form, buttonStyling } from "@/components/ui"
+import { buttonStyling } from "@/components/ui"
 import { user } from "@/lib/db"
 import { alreadyFilledIn } from "@/lib/app-api"
 import { base64EncodeFile, postJsonData } from "@/lib/util/request"
 import { DataType as DetailsApiDataType } from "@/pages/api/details"
 import { DoneeInfo } from "@/types/db"
+import { Fieldset, ImageInput, Legend, TextInput } from "@/components/form"
 
 type Props = {
   doneeInfo: DoneeInfo
@@ -73,9 +73,9 @@ export default function Details({ doneeInfo, itemsFilledIn }: Props) {
 
   return (
     <form ref={formRef} onSubmit={onSubmit} className="w-full max-w-2xl space-y-4 p-4">
-      <Form.Fieldset className="grid gap-4 sm:grid-cols-2 sm:gap-6">
-        <Form.Legend className="sm:col-span-2">Organisation</Form.Legend>
-        <Form.TextInput
+      <Fieldset className="grid gap-4 sm:grid-cols-2 sm:gap-6">
+        <Legend className="sm:col-span-2">Organisation</Legend>
+        <TextInput
           id="companyAddress"
           minLength={10}
           defaultValue={doneeInfo.companyAddress}
@@ -83,40 +83,40 @@ export default function Details({ doneeInfo, itemsFilledIn }: Props) {
           className="sm:col-span-2"
           required
         />
-        <Form.TextInput
+        <TextInput
           id="companyName"
           defaultValue={doneeInfo.companyName}
           label="Legal name"
           required
         />
-        <Form.TextInput
+        <TextInput
           id="country"
           minLength={2}
           defaultValue={doneeInfo.country}
           label="Country"
           required
         />
-        <Form.TextInput
+        <TextInput
           id="registrationNumber"
           minLength={15}
           defaultValue={doneeInfo.registrationNumber}
           label="Charity registration number"
           required
         />
-        <Form.TextInput
+        <TextInput
           id="signatoryName"
           minLength={5}
           label="Signatory's name"
           defaultValue={doneeInfo.signatoryName}
           required
         />
-        <Form.ImageInput
+        <ImageInput
           id="signature"
           label="Image of signatory's signature"
           helper={doneeInfo.signatoryName ? imageNotRequiredHelper : imageHelper}
           required={!Boolean(doneeInfo.signatoryName)}
         />
-        <Form.ImageInput
+        <ImageInput
           id="smallLogo"
           label="Small image of organisation's logo"
           helper={doneeInfo.smallLogo ? imageNotRequiredHelper : imageHelper}
@@ -127,7 +127,7 @@ export default function Details({ doneeInfo, itemsFilledIn }: Props) {
           type="submit"
           value={itemsFilledIn ? "Generate Receipts" : "Select Qualifying Items"}
         />
-      </Form.Fieldset>
+      </Fieldset>
     </form>
   )
 }
